+++ packages/rum/src/boot/startRecording.ts
@@ -1,4 +1,4 @@
-import { timeStampNow } from '@datadog/browser-core'
+import { timeStampNow, createHttpRequest } from '@datadog/browser-core'
 import type {
   LifeCycle,
   ViewContexts,
@@ -10,7 +10,7 @@ import { LifeCycleEventType } from '@datadog/browser-rum-core'
 
 import { record } from '../domain/record'
 import type { DeflateWorker } from '../domain/segmentCollection'
-import { startSegmentCollection } from '../domain/segmentCollection'
+import { startSegmentCollection, SEGMENT_BYTES_LIMIT } from '../domain/segmentCollection'
 import { send } from '../transport/send'
 import { RecordType } from '../types'
 
@@ -19,21 +19,21 @@ export function startRecording(
   configuration: RumConfiguration,
   sessionManager: RumSessionManager,
   viewContexts: ViewContexts,
-  worker: DeflateWorker
+  worker: DeflateWorker,
+  httpRequest = createHttpRequest(configuration.sessionReplayEndpointBuilder, SEGMENT_BYTES_LIMIT)
 ) {
   const { addRecord, stop: stopSegmentCollection } = startSegmentCollection(
     lifeCycle,
     configuration.applicationId,
     sessionManager,
     viewContexts,
-    (data, metadata, rawSegmentBytesCount) =>
-      send(configuration.sessionReplayEndpointBuilder, data, metadata, rawSegmentBytesCount),
+    (data, metadata, rawSegmentBytesCount) => send(httpRequest, data, metadata, rawSegmentBytesCount),
     worker
   )
 
   const {
     stop: stopRecording,
-    takeFullSnapshot,
+    takeSubsequentFullSnapshot,
     flushMutations,
   } = record({
     emit: addRecord,
@@ -51,7 +51,7 @@ export function startRecording(
   const { unsubscribe: unsubscribeViewCreated } = lifeCycle.subscribe(
     LifeCycleEventType.VIEW_CREATED,
     (view: ViewCreatedEvent) => {
-      takeFullSnapshot(view.startClocks.timeStamp)
+      takeSubsequentFullSnapshot(view.startClocks.timeStamp)
     }
   )